Move enumeration logic to iterators
